LayoutTests/canvas/philip/tests/2d.imageData.create1.zero-expected.txt

< [index] >
2d.imageData.create1.zero
createImageData(null) throws INDEX_SIZE_ERR
References: 2d.imageData.create.null
Actual output:
Passed
